backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.0
Django==5.1.3
django-environ==0.11.2
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
